Fix: Stop causing Anki to "forget" done reviews. Writes use flush(). Using flush() changes the last modified time of the card. This can cause problems when you review cards on one device, and reset ease on another. Only one of this changes will "survive" because Anki doesn't merge changes within individual cards. It will just take the lastest edit.
Hence the change. If the ease factor of a card is only reset after
it has been changed (after a review) it either:
was reviewed on the same device and the review has to be uploaded anyway
...or it was synched from another device. In this
case it doesn't overwrite
the changes from the other device, but applies them on top.